feat: add gget nonce to nre and cli
