+++ src/nile/core/account.py
@@ -5,9 +5,9 @@ import os
 from dotenv import load_dotenv
 
 from nile import accounts, deployments
-from nile.common import get_nonce
 from nile.core.call_or_invoke import call_or_invoke
 from nile.core.deploy import deploy
+from nile.utils.get_nonce import get_nonce
 
 try:
     from nile.signer import Signer
