+++ src/nile/nre.py
@@ -7,6 +7,7 @@ from nile.core.declare import declare
 from nile.core.deploy import deploy
 from nile.core.plugins import get_installed_plugins, skip_click_exit
 from nile.utils.get_accounts import get_accounts
+from nile.utils.get_nonce import get_nonce as get_nonce_command
 
 
 class NileRuntimeEnvironment:
@@ -55,3 +56,7 @@ class NileRuntimeEnvironment:
     def get_accounts(self):
         """Retrieve and manage deployed accounts."""
         return get_accounts(self.network)
+
+    def get_nonce(self, contract_address):
+        """Retrieve the nonce for a contract."""
+        return get_nonce_command(contract_address, self.network)
